Introduce LiteralExpression
